workaround: patch openapi spec generator
patch spec generation, allowing the rust api client to be generated out of the box.

update types + codegen

0.20 rebased

+++ src/ingester/parser/state_update.rs
@@ -24,7 +24,7 @@ pub struct EnrichedPathNode {
     pub seq: u64,
     pub level: usize,
     pub tree_depth: usize,
-    pub leaf_index: Option<u32>,
+    pub leaf_index: Option<u64>,
 }
 
 pub struct PathUpdate {
